# openapi_specification.yaml
openapi: 3.0.3
info:
  title: «Дәресханә» (Tatar OCR) — API Contract Specification
  version: 1.0.0
  description: |
    This document provides the definitive API contracts and architectural design for **«Дәресханә»**, built around two foundational principles:
    1. **Zero-Photo Transmission**: Camera images never leave the teacher's phone. All ArUco perspective rectifications, cell cropping, and TFLite neural network character recognitions happen 100% locally on the device in memory.
    2. **Backend Analytical Processing**: The phone uploads only lightweight structured verification results (1–2 KB JSON per sheet). The backend performs all heavy historical data aggregation, student learning curve analysis, class heatmaps, error categorizations, and gradebook exports.
servers:
  - url: /
    description: Base Server URL
components:
  securitySchemes:
    TeacherUUID:
      type: apiKey
      in: header
      name: X-Teacher-UUID
      description: Teacher UUID v4 generated upon first install
  parameters:
...
    HeaderTeacherUUID:
      name: X-Teacher-UUID
      in: header
      required: false
      description: Teacher UUID v4
      schema:
        type: string
        format: uuid
        example: "c56a4180-65aa-42ec-a945-5fd21dec0538"
  schemas:
...
    AssignmentMetadataResponse:
      type: object
      description: Metadata, title, grade level, topic tags, points.
      properties:
        assignment_id:
          type: string
          example: "TAT-2026-Q1"
        title:
          type: string
          example: "Татар теле. 7 сыйныф. Исем килешләре"
        grade_level:
          type: integer
          example: 7
        topic_tags:
          type: array
          items:
            type: string
          example: ["case_dative", "case_ablative"]
        points:
          type: number
          example: 8.0
...
    FrequentWeakTopic:
      type: object
      properties:
        topic_code:
          type: string
          example: "case_locative"
        topic_name_tt:
          type: string
          example: "Урын-вакыт килеше (-да/-дә, -та/-тә)"
        total_questions:
          type: integer
          example: 10
        wrong_count:
          type: integer
          example: 4
        accuracy_pct:
          type: number
          format: float
          example: 60.0
    ProblematicLetter:
      type: object
      properties:
        letter:
          type: string
          example: "Ң"
        total_occurrences:
          type: integer
          example: 14
        misrecognized_or_wrong:
          type: integer
          example: 4
        accuracy_pct:
          type: number
          format: float
          example: 71.4
        common_confusions:
          type: array
          items:
            type: string
          example: ["Н"]
    StudentHistoryItem:
      type: object
      properties:
        submission_id:
          type: string
          example: "sub_8f23a"
        assignment_id:
          type: string
          example: "TAT-2026-Q1"
        assignment_title:
          type: string
          example: "Исем килешләре"
        date:
          type: string
          format: date-time
          example: "2026-09-17T09:30:00Z"
        score:
          type: number
          example: 7
        max_score:
          type: number
          example: 8
        grade:
          type: integer
          example: 4
    StudentAnalyticsResponse:
      type: object
      properties:
        student_id:
          type: string
          example: "stu_01"
        full_name:
          type: string
          example: "Галиев Амир Рустемович"
        class_name:
          type: string
          example: "7-А"
        total_tests_completed:
          type: integer
          example: 12
        average_score_pct:
          type: number
          format: float
          example: 82.5
        average_grade:
          type: number
          format: float
          example: 4.3
        grade_distribution:
          type: object
          additionalProperties:
            type: integer
          example:
            "5": 5
            "4": 5
            "3": 2
            "2": 0
        frequent_weak_topics:
          type: array
          items:
            $ref: '#/components/schemas/FrequentWeakTopic'
        problematic_letters:
          type: array
          items:
            $ref: '#/components/schemas/ProblematicLetter'
        history:
          type: array
          items:
            $ref: '#/components/schemas/StudentHistoryItem'
    TopClassMistake:
      type: object
      properties:
        topic_code:
          type: string
          example: "case_ablative"
        topic_name_tt:
          type: string
          example: "Чыгыш килешендә сингармонизм ялгышлары (-дан/-дән, -тан/-тән)"
        failure_rate_pct:
          type: number
          format: float
          example: 42.3
        affected_students_count:
          type: integer
          example: 11
    DifficultCharacter:
      type: object
      properties:
        letter:
          type: string
          example: "Ң"
        error_rate_pct:
          type: number
          format: float
          example: 36.5
    StudentPerformanceRow:
      type: object
      properties:
        student_id:
          type: string
          example: "stu_02"
        full_name:
          type: string
          example: "Закирова Ләйсән Ильдаровна"
        average_score_pct:
          type: number
          format: float
          example: 94.0
        average_grade:
          type: number
          format: float
          example: 4.8
        tests_completed:
          type: integer
          example: 8
    ClassAnalyticsResponse:
      type: object
      properties:
        class_id:
          type: string
          example: "cls_7a_2026"
        class_name:
          type: string
          example: "7-А"
        students_count:
          type: integer
          example: 25
        average_class_score_pct:
          type: number
          format: float
          example: 81.2
        grade_distribution:
          type: object
          additionalProperties:
            type: integer
          example:
            "5": 9
            "4": 11
            "3": 4
            "2": 1
        top_class_mistakes:
          type: array
          items:
            $ref: '#/components/schemas/TopClassMistake'
        difficult_characters_across_class:
          type: array
          items:
            $ref: '#/components/schemas/DifficultCharacter'
        students_performance_table:
          type: array
          items:
            $ref: '#/components/schemas/StudentPerformanceRow'
    WrongSubmissionItem:
      type: object
      properties:
        student_id:
          type: string
          example: "stu_03"
        student_name:
          type: string
          example: "Хабибуллин Тимур"
        written_answer:
          type: string
          example: "КИТАПГА"
    QuestionAnalyticsItem:
      type: object
      properties:
        question_number:
          type: integer
          example: 1
        marker_id:
          type: integer
          example: 11
        prompt:
          type: string
          example: "Куегыз сүзне юнәлеш килешендә: китап ->"
        expected_answer:
          type: string
          example: "КИТАПКА"
        accuracy_pct:
          type: number
          format: float
          example: 92.0
        wrong_submissions:
          type: array
          items:
            $ref: '#/components/schemas/WrongSubmissionItem'
    AssignmentAnalyticsResponse:
      type: object
      properties:
        assignment_id:
          type: string
          example: "TAT-2026-Q1"
        title:
          type: string
          example: "Исем килешләре һәм кушымчалар"
        total_submissions:
          type: integer
          example: 25
        average_score_pct:
          type: number
          format: float
          example: 83.2
        questions_analytics:
          type: array
          items:
            $ref: '#/components/schemas/QuestionAnalyticsItem'
paths:
...
  /api/v1/assignments/{assignment_id}:
    get:
      tags:
        - Assignment Delivery
      summary: Get Assignment Metadata
      description: Metadata, title, grade level, topic tags, points.
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
          example: "TAT-2026-Q1"
      responses:
        '200':
          description: Metadata, title, grade level, topic tags, points.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentMetadataResponse'
...
  /api/v1/analytics/students/{student_id}:
    get:
      tags:
        - Backend Analytics Engine
      summary: Student Analytics (Individual)
      description: Computes learning curves, error frequencies by topic, and specific letter confusions for an individual student.
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
          example: "stu_01"
        - $ref: '#/components/parameters/HeaderTeacherUUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAnalyticsResponse'
  /api/v1/analytics/classes/{class_id}:
    get:
      tags:
        - Backend Analytics Engine
      summary: Class Aggregated Analytics & Student Leaderboard
      description: Class heatmap identifies which grammar rules the teacher must re-explain on the next lesson, plus a full pupil performance table.
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
          example: "cls_7a_2026"
        - $ref: '#/components/parameters/HeaderTeacherUUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassAnalyticsResponse'
  /api/v1/analytics/assignments/{assignment_id}:
    get:
      tags:
        - Backend Analytics Engine
      summary: Assignment Analytics
      description: Item analysis per question across all completed sheets.
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            type: string
          example: "TAT-2026-Q1"
        - $ref: '#/components/parameters/HeaderTeacherUUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentAnalyticsResponse'
...